pyproject.toml
[tool.poetry]
name = "relais"
version = "0.1.0"
description = "RELAIS Micro-brick Architecture"
authors = ["Benjamin Marchand"]
readme = "README.md"
packages = [{include = "common"}, {include = "relay"}, {include = "gateway"}, {include = "workshop"}, {include = "sentinel"}, {include = "memory"}, {include = "archivist"}]

[tool.poetry.dependencies]
python = ">=3.11"
redis = "^5.0.0"
litellm = "^1.25.0"
supervisor = "^4.2.5"
httpx = ">=0.27"
aiosqlite = ">=0.20"
pyyaml = ">=6.0"
python-dotenv = ">=1.0"
pydantic = ">=2.9"
anthropic = ">=0.40"
sqlmodel = ">=0.14"
alembic = ">=1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = [
    "unit: fast, isolated unit tests",
    "integration: tests that require external services (Redis, LiteLLM, etc.)",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "sqlmodel>=0.0.37",
]
